third_party/blink/web_tests/http/tests/security/local-video-src-from-remote-expected.txt

CONSOLE ERROR: Not allowed to load local resource: test.mp4
Test that a remote video element can not use a local 'src'

This test only works in DRT

EVENT(error)

SUCCESS: local movie failed to load

END OF TEST
